move prim_utils into separate file
